--- frontend/src/components/LocationCarousel.jsx
@@ -1,13 +1,0 @@
-import Carousel from "react-bootstrap/Carousel";
-
-export default function LocationCarousel() {
-  return (
-    <Carousel.Item>
-      {/* <ExampleCarouselImage text="First slide" /> */}
-      <Carousel.Caption>
-        <h3>First slide label</h3>
-        <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>
-      </Carousel.Caption>
-    </Carousel.Item>
-  );
-}
