--- frontend/src/components/LoginModal.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function LoginModal() {
-  return <div>LoginModal</div>;
-}
