--- frontend/src/contexts/AppContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const AppContext = createContext(null);
-
-export default AppContext;
