+++ frontend/src/index.js
@@ -1,13 +1,10 @@
 import ReactDOM from "react-dom/client";
-import { RouterProvider } from "react-router-dom";
-import router from "./Router";
-// import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import "bootstrap/dist/css/bootstrap.min.css";
-import App2 from "./App2";
+import "./index.css";
 import App from "./App";
 
-ReactDOM.createRoot(document.getElementById("root")).render(<App2 />);
+ReactDOM.createRoot(document.getElementById("root")).render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
