+++ frontend/src/modals/ContactUs.jsx
@@ -1,0 +1,52 @@
+import { useState } from "react";
+import Button from "react-bootstrap/Button";
+import Modal from "react-bootstrap/Modal";
+
+export default function ContactUs() {
+  const [show, setShow] = useState(false);
+
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
+
+  return (
+    <>
+      <Button className="bg-secondary border-secondary" onClick={handleShow}>
+        ContactUs
+      </Button>
+
+      <Modal
+        show={show}
+        onHide={handleClose}
+        backdrop="static"
+        keyboard={false}
+      >
+        <Modal.Header closeButton>
+          <Modal.Title>ContactUs</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <p>info@ehra.com </p> <p> 800-GET-EHRA</p>
+          <p>
+            Pellentesque habitant morbi tristique senectus et netus et malesuada
+            fames ac turpis egestas. Vestibulum tortor quam, feugiat vitae,
+            ultricies eget, tempor sit amet, ante. Donec eu libero sit amet quam
+            egestas semper. Aenean ultricies mi vitae est. Mauris placerat
+            eleifend leo. Quisque sit amet est et sapien ullamcorper pharetra.
+            Vestibulum erat wisi, condimentum sed, commodo vitae, ornare sit
+            amet, wisi. Aenean fermentum, elit eget tincidunt condimentum, eros
+            ipsum rutrum orci, sagittis tempus lacus enim ac dui. Donec non enim
+            in turpis pulvinar facilisis. Ut felis. Praesent dapibus, neque id
+            cursus faucibus, tortor neque egestas augue, eu vulputate magna eros
+            eu erat. Aliquam erat volutpat. Nam dui mi, tincidunt quis, accumsan
+            porttitor, facilisis luctus, metus
+          </p>
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleClose}>
+            Close
+          </Button>
+          <Button variant="primary">Understood</Button>
+        </Modal.Footer>
+      </Modal>
+    </>
+  );
+}
