--- frontend/src/pages/ProviderPage.jsx
@@ -1,6 +1,0 @@
-import { useParams } from "react-router-dom";
-
-export default function ProviderPage() {
-  const { id } = useParams();
-  return <div>{id}</div>;
-}
