--- frontend/src/pages/SignUpPage.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function SignUpPage() {
-  return <div>SignUpPage</div>;
-}
