--- frontend/src/pages/UserPage.jsx
@@ -1,6 +1,0 @@
-import { useParams } from "react-router-dom";
-
-export default function UserPage() {
-  const { id } = useParams();
-  return <div>{id}</div>;
-}
